Divide in two packages
- Divide in packages uri and spiffe
- Use explicit return values
