+++ Makefile
@@ -1,5 +1,12 @@
+ROOT := github.com/spiffe/go-spiffe
+PACKAGE_LIST := uri spiffe
+
 test: vet
-	go test
+	@for p in $(PACKAGE_LIST); do \
+		go test $(ROOT)/$$p || exit 1; \
+	done
 
 vet:
-	go vet
+	@for p in $(PACKAGE_LIST); do \
+		go vet $(ROOT)/$$p; \
+	done
