+++ uri/uri_test.go
@@ -1,16 +1,12 @@
-package spiffe_test
+package uri
 
 import (
-	"crypto/rand"
-	"crypto/rsa"
-	"crypto/x509"
 	"crypto/x509/pkix"
 	"io/ioutil"
-	"math/big"
 	"strings"
 	"testing"
 
-	spiffe "github.com/spiffe/go-spiffe"
+	"github.com/spiffe/go-spiffe/spiffe"
 )
 
 func getCertificateFromFile(t *testing.T, certFilePath string) string {
@@ -25,9 +21,9 @@ func getCertificateFromFile(t *testing.T, certFilePath string) string {
 const golden = "spiffe://dev.acme.com/path/service"
 
 func TestGetURINamesFromPEM(t *testing.T) {
-	certPEM := getCertificateFromFile(t, "testdata/leaf.cert.pem")
+	certPEM := getCertificateFromFile(t, "../testdata/leaf.cert.pem")
 
-	uris, err := spiffe.GetURINamesFromPEM(string(certPEM))
+	uris, err := GetURINamesFromPEM(string(certPEM))
 	if err != nil {
 		t.Error(err)
 	}
@@ -40,8 +36,8 @@ func TestGetURINamesFromPEM(t *testing.T) {
 		t.Fatalf("Expected 1 URI but got '%v'", len(uris))
 	}
 
-	certPEM = getCertificateFromFile(t, "testdata/intermediate.cert.pem")
-	uris, err = spiffe.GetURINamesFromPEM(string(certPEM))
+	certPEM = getCertificateFromFile(t, "../testdata/intermediate.cert.pem")
+	uris, err = GetURINamesFromPEM(string(certPEM))
 	if err == nil {
 		t.Fatal("Expected to fail")
 	}
@@ -119,7 +115,7 @@ fDGmAkjHrspiorTruphHk+cymJfjAGqZ0l6il4Wi5w4R5jrxnJMhkxbjr/PG5xCS
 )
 
 func TestFGetURINamesFromPEM(t *testing.T) {
-	uris, err := spiffe.FGetURINamesFromPEM(strings.NewReader(goodCert))
+	uris, err := FGetURINamesFromPEM(strings.NewReader(goodCert))
 	if err != nil {
 		t.Error(err)
 	}
@@ -132,7 +128,7 @@ func TestFGetURINamesFromPEM(t *testing.T) {
 		t.Fatalf("Expected 1 URI but got '%v'", len(uris))
 	}
 
-	uris, err = spiffe.FGetURINamesFromPEM(strings.NewReader(badCert))
+	uris, err = FGetURINamesFromPEM(strings.NewReader(badCert))
 	if err == nil {
 		t.Fatal("Expected to fail")
 	}
@@ -143,7 +139,7 @@ func TestFGetURINamesFromPEM(t *testing.T) {
 }
 
 func TestMarshalUriSANsAndGetURINamesFromExtensions(t *testing.T) {
-	uriSans, err := spiffe.MarshalUriSANs([]string{golden})
+	uriSans, err := MarshalUriSANs([]string{golden})
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -161,7 +157,7 @@ func TestMarshalUriSANsAndGetURINamesFromExtensions(t *testing.T) {
 		},
 	}
 
-	uris, err := spiffe.GetURINamesFromExtensions(&extensions)
+	uris, err := GetURINamesFromExtensions(&extensions)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -171,39 +167,3 @@ func TestMarshalUriSANsAndGetURINamesFromExtensions(t *testing.T) {
 	}
 }
 
-func TestGetKeyUsageExtensionsFromCertificate(t *testing.T) {
-	key, err := rsa.GenerateKey(rand.Reader, 2048)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	serial, err := rand.Int(rand.Reader, new(big.Int).Lsh(big.NewInt(1), 128))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	template := &x509.Certificate{
-		SerialNumber: serial,
-		KeyUsage: x509.KeyUsageKeyEncipherment |
-			x509.KeyUsageDigitalSignature |
-			x509.KeyUsageCertSign,
-		BasicConstraintsValid: true,
-		IsCA: true,
-	}
-
-	der, err := x509.CreateCertificate(rand.Reader,
-		template, template, &key.PublicKey, key)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	cert, err := x509.ParseCertificate(der)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	keyUsageExtensions := spiffe.GetKeyUsageExtensionsFromCertificate(cert)
-	if len(keyUsageExtensions) != 1 {
-		t.Fatalf("Expected to get one Key Usage extension, but got: %v", len(keyUsageExtensions))
-	}
-}
